refactor: extract getArticleKeys and loadSessionArticles to shared session-utils
Extract duplicated article loading logic and identifier key extraction
into src/cli/commands/session-utils.ts, replacing inline implementations
in export, summary, diff, and register commands.
